requirements.txt
Flask==0.10.1
Flask-Bcrypt==0.6.0
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-SQLAlchemy==1.0
Flask-Uploads==0.1.3
Flask-WTF==0.9.5
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.1
SQLAlchemy==0.9.6
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.6
blinker==1.3
boto==2.32.1
decorator==3.4.0
gunicorn==19.1.0
itsdangerous==0.24
pbr==0.8.2
psycopg2==2.5.3
py-bcrypt==0.4
requests==2.3.0
six==1.7.3
sqlalchemy-migrate==0.9.1
wsgiref==0.1.2
